refactor(cooldowns): sidebar editation
