+++ locales/en/ui.systems.cooldown.json
@@ -3,9 +3,6 @@
   "empty": "No cooldowns were created yet.",
   "emptyAfterSearch": "No cooldown were found by your search for <strong>\"$search\"</strong>.",
   "notify-as-whisper": "Notify as whisper",
-  "error": {
-    "isEmpty": "This value cannot be empty"
-  },
   "settings": {
     "enabled": "Status",
     "cooldownNotifyAsWhisper": "Whisper cooldown informations",
